+++ app/f1/my-plans/page.tsx
@@ -1,0 +1,7 @@
+import { MyPlansClient } from "./MyPlansClient";
+
+export const dynamic = "force-dynamic";
+
+export default function MyPlansPage() {
+  return <MyPlansClient />;
+}